fixed stupid stupid bug, added SPW support to CP.
git-svn-id: file:///home/svn/omero/trunk@7363 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/tools/cellprofiler/CPsubfunctions/CPOMEROimread.m
@@ -35,7 +35,7 @@ elseif nargin == 4,
     handles = varargin{4};
     [pixelsId, z, t] = parseFileDetails(CurrentFileName);
 
-    rawPlane = omeroGateway.getPlane(pixelsId, z, c, t);
+    rawPlane = omeroGateway.getPlane(int32(pixelsId), int32(z), int32(c), int32(t));
     pixels = omeroGateway.getPixels(pixelsId);
 
     pixelType = pixels.getPixelsType.getValue.getValue.toCharArray;
